refactor(provider-inference): move provider tests closer to implementations
- Move azure-openai and openrouter-ai test files from stage-ui to provider-inference
- Update test imports to use direct provider definitions instead of getDefinedProvider()
- Add @proj-airi/core-agent as devDependency for test utilities
