+++ packages/provider-inference/package.json
@@ -53,6 +53,7 @@
     "zod": "catalog:"
   },
   "devDependencies": {
+    "@proj-airi/core-agent": "workspace:^",
     "@types/node": "catalog:",
     "@vitest/browser-playwright": "catalog:vitest",
     "@xsai/generate-speech": "catalog:",
